requirements.txt
numpy==2.1.1
opencv-python==4.10.0.84
pillow==10.4.0
python-poppler==0.4.1
typing_extensions==4.12.2
